requirements.txt
alembic==1.2.1
Babel==2.7.0
boto3==1.9.246
botocore==1.12.246
Click==7.0
docutils==0.15.2
dominate==2.4.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
